pyproject.toml
[tool.poetry]
name = "blanketcon-cc"
version = "0.1.0"
description = ""
authors = ["Jonathan Coates <git@squiddev.cc>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
